Chat_With_multiple_Pdf_Documents/requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain_google_genai
##with help of langchain we can acces api of google ai models